requirements.txt
aiohttp==3.1.3
asyncpg==0.15.0
marshmallow==2.13.5
tortoise-orm==0.7.4
uvloop==0.9.1
PyYAML==3.12